Extract duplicate code to seperate file
