+++ web_src/js/utils.js
@@ -139,3 +139,5 @@ export function parseDom(text, contentType) {
 export function serializeXml(node) {
   return xmlSerializer.serializeToString(node);
 }
+
+export const sleep = ms => new Promise(resolve => setTimeout(resolve, ms))
